rename addscore.js to modules.js and delete useless export from index.js
